More tests

+++ src/test/java/model/TestBank.java
@@ -39,7 +39,7 @@ public class TestBank {
 	public void testGetId() {
 		initBank();
 		tested.setId(1);
-		assertEquals(1,tested.getId());
+		assertEquals(1,tested.getId().intValue());
 	}
 	@Test(expected = IllegalArgumentException.class)
 	public void testSetId_Invalid() {
